Remove Azure OpenAI support completely
Remove all Azure OpenAI integration from the rest-narration service,
leaving OpenAI as the sole AI provider. The LangChain4J abstraction
keeps the business logic provider-agnostic, so no service code changes
were needed.

- Remove quarkus-langchain4j-azure-openai dependency
- Remove all Azure OpenAI configuration and profiles from
  application.properties
- Collapse the test class hierarchy: merge the abstract
  NarrationResourceIT base class into the sole remaining subclass,
  now a standalone NarrationResourceIT
- Remove Azure OpenAI WireMock stubs from narration.json
- Simplify CI matrix: convert object-format entries back to plain
  strings in simple-build-test.yml, remove openai-type from
  build-push-container-images.yml and compute-container-tag.sh,
  simplify calculateModules.java
- Delete Azure deployment scripts, docs, and architecture diagrams
- Update all documentation to remove Azure OpenAI references
