+++ .github/workflows/build-push-container-images.yml
@@ -121,7 +121,7 @@ jobs:
         arch:
           - amd64
           - arm64
-    name: "Build JVM images (${{ matrix.arch }}-${{ matrix.project.name }}-java${{ matrix.java }}-${{ matrix.project.openai-type }})"
+    name: "Build JVM images (${{ matrix.arch }}-${{ matrix.project.name }}-java${{ matrix.java }})"
     steps:
       - name: Checkout from ${{ needs.check-commit-count.outputs.ref }}
         uses: actions/checkout@v6
@@ -144,14 +144,9 @@ jobs:
             echo "JVM_DOCKERFILE=Dockerfile.jvm${{ matrix.java }}" >> "$GITHUB_ENV"
           fi
 
-      - name: Create OpenAI env vars
-        if: matrix.project.openai-type
-        working-directory: ${{ matrix.project.name }}
-        run: echo "QUARKUS_PROFILE=${{ matrix.project.openai-type }}" >> $GITHUB_ENV
-
       - name: Create CONTAINER_TAG
         working-directory: .github/workflows/scripts
-        run: echo "CONTAINER_TAG=$(./compute-container-tag.sh "" ${{ matrix.java }} ${{ env.LATEST_IMAGE_TAG }} ${{ needs.check-commit-count.outputs.branch }} "${{ matrix.project.openai-type }}")" >> $GITHUB_ENV
+        run: echo "CONTAINER_TAG=$(./compute-container-tag.sh "" ${{ matrix.java }} ${{ env.LATEST_IMAGE_TAG }} ${{ needs.check-commit-count.outputs.branch }})" >> $GITHUB_ENV
 
       - name: Set up QEMU
         if: matrix.arch == 'arm64'
@@ -164,7 +159,7 @@ jobs:
         with:
           install: true
 
-      - name: Build JVM image (${{ matrix.project.name }}-java${{ matrix.java }}-${{ matrix.arch }}-${{ matrix.project.openai-type }})
+      - name: Build JVM image (${{ matrix.project.name }}-java${{ matrix.java }}-${{ matrix.arch }})
         working-directory: ${{ matrix.project.name }}
         run: |
           ./mvnw -B clean package -DskipTests \
@@ -176,7 +171,7 @@ jobs:
             -Dquarkus.docker.dockerfile-jvm-path=src/main/docker/${{ env.JVM_DOCKERFILE }} \
             -Dquarkus.docker.buildx.platform=linux/${{ matrix.arch }}
 
-      - name: Save JVM Image (${{ matrix.project.name }}-java${{ matrix.java }}-${{ matrix.arch }}-${{ matrix.project.openai-type }})
+      - name: Save JVM Image (${{ matrix.project.name }}-java${{ matrix.java }}-${{ matrix.arch }})
         uses: ./.github/actions/save-container-image
         with:
           image: "${{ env.IMAGE_BASE_NAME }}/${{ matrix.project.name }}:${{ env.CONTAINER_TAG }}-${{ matrix.arch }}"
@@ -196,7 +191,7 @@ jobs:
         arch:
           - amd64
           - arm64
-    name: "Build Native images (${{ matrix.arch }}-${{ matrix.project.name }}-java${{ matrix.java.jvm }}-${{ matrix.project.openai-type }})"
+    name: "Build Native images (${{ matrix.arch }}-${{ matrix.project.name }}-java${{ matrix.java.jvm }})"
     steps:
       - name: Checkout from ${{ needs.check-commit-count.outputs.ref }}
         uses: actions/checkout@v6
@@ -210,14 +205,9 @@ jobs:
           distribution: temurin
           cache: maven
 
-      - name: Create OpenAI env vars
-        if: matrix.project.openai-type
-        working-directory: ${{ matrix.project.name }}
-        run: echo "QUARKUS_PROFILE=${{ matrix.project.openai-type }}" >> $GITHUB_ENV
-
       - name: Create CONTAINER_TAG
         working-directory: .github/workflows/scripts
-        run: echo "CONTAINER_TAG=$(./compute-container-tag.sh "native-" ${{ matrix.java.jvm }} ${{ env.LATEST_IMAGE_TAG }} ${{ needs.check-commit-count.outputs.branch }} "${{ matrix.project.openai-type }}")" >> $GITHUB_ENV
+        run: echo "CONTAINER_TAG=$(./compute-container-tag.sh "native-" ${{ matrix.java.jvm }} ${{ env.LATEST_IMAGE_TAG }} ${{ needs.check-commit-count.outputs.branch }})" >> $GITHUB_ENV
 
       - name: Set up QEMU
         if: matrix.arch == 'arm64'
@@ -230,7 +220,7 @@ jobs:
         with:
           install: true
 
-      - name: Build native image (${{ matrix.project.name }}-java${{ matrix.java.jvm }}-${{ matrix.arch }}-${{ matrix.project.openai-type }})
+      - name: Build native image (${{ matrix.project.name }}-java${{ matrix.java.jvm }}-${{ matrix.arch }})
         working-directory: ${{ matrix.project.name }}
         run: |
           ./mvnw -B clean package -DskipTests -Pnative \
@@ -244,7 +234,7 @@ jobs:
             -Dquarkus.container-image.tag=${{ env.CONTAINER_TAG }}-${{ matrix.arch }} \
             -Dquarkus.docker.buildx.platform=linux/${{ matrix.arch }}
 
-      - name: Save native Image (${{ matrix.project.name }}-java${{ matrix.java.jvm }}-${{ matrix.arch }}-${{ matrix.project.openai-type }})
+      - name: Save native Image (${{ matrix.project.name }}-java${{ matrix.java.jvm }}-${{ matrix.arch }})
         uses: ./.github/actions/save-container-image
         with:
           image: "${{ env.IMAGE_BASE_NAME }}/${{ matrix.project.name }}:${{ env.CONTAINER_TAG }}-${{ matrix.arch }}"
@@ -268,7 +258,7 @@ jobs:
         arch:
           - amd64
           - arm64
-    name: "Push app images (${{ matrix.arch }}-${{ matrix.project.name }}-${{ matrix.kind }}java${{ matrix.java }}-${{ matrix.project.openai-type }})"
+    name: "Push app images (${{ matrix.arch }}-${{ matrix.project.name }}-${{ matrix.kind }}java${{ matrix.java }})"
     steps:
       - name: Checkout from ${{ needs.check-commit-count.outputs.ref }}
         uses: actions/checkout@v6
@@ -277,7 +267,7 @@ jobs:
 
       - name: Create CONTAINER_TAG
         working-directory: .github/workflows/scripts
-        run: echo "CONTAINER_TAG=$(./compute-container-tag.sh "${{ matrix.kind }}" ${{ matrix.java }} ${{ env.LATEST_IMAGE_TAG }} ${{ needs.check-commit-count.outputs.branch }} "${{ matrix.project.openai-type }}")" >> $GITHUB_ENV
+        run: echo "CONTAINER_TAG=$(./compute-container-tag.sh "${{ matrix.kind }}" ${{ matrix.java }} ${{ env.LATEST_IMAGE_TAG }} ${{ needs.check-commit-count.outputs.branch }})" >> $GITHUB_ENV
 
       - name: Get saved images (${{ matrix.project.name }}:${{ env.CONTAINER_TAG }}-${{ matrix.arch }})
         uses: ./.github/actions/load-container-image
@@ -310,7 +300,7 @@ jobs:
           - ""
           - "native-"
         project: ${{ fromJson(needs.calculate-modules.outputs.projects_matrix) }}
-    name: Create app multiarch manifests (${{ matrix.project.name }}-${{ matrix.kind }}java${{ matrix.java }}-${{ matrix.project.openai-type }})
+    name: Create app multiarch manifests (${{ matrix.project.name }}-${{ matrix.kind }}java${{ matrix.java }})
     steps:
       - name: Checkout from ${{ needs.check-commit-count.outputs.ref }}
         uses: actions/checkout@v6
@@ -319,7 +309,7 @@ jobs:
 
       - name: Create CONTAINER_TAG
         working-directory: .github/workflows/scripts
-        run: echo "CONTAINER_TAG=$(./compute-container-tag.sh "${{ matrix.kind }}" ${{ matrix.java }} ${{ env.LATEST_IMAGE_TAG }} ${{ needs.check-commit-count.outputs.branch }} "${{ matrix.project.openai-type }}")" >> $GITHUB_ENV
+        run: echo "CONTAINER_TAG=$(./compute-container-tag.sh "${{ matrix.kind }}" ${{ matrix.java }} ${{ env.LATEST_IMAGE_TAG }} ${{ needs.check-commit-count.outputs.branch }})" >> $GITHUB_ENV
 
       - name: Login to quay
         uses: docker/login-action@v4.2.0
