+++ images/plantuml/application-architecture.puml
@@ -43,7 +43,7 @@ node "Narration" as narration {
   narrationQuarkus .up> langchain
 }
 
-cloud "OpenAI\nAzure OpenAI" as openai
+cloud "OpenAI" as openai
 
 node "Hero" as hero {
     agent "Quarkus (reactive)" <<application>> as heroQuarkus
