+++ rest-narration/src/main/resources/application.properties
@@ -17,45 +17,20 @@ quarkus.langchain4j.openai.max-retries=3
 quarkus.langchain4j.openai.chat-model.temperature=1
 quarkus.langchain4j.openai.chat-model.top-p=0.5
 
-## Default Azure OpenAI configuration
-quarkus.langchain4j.azure-openai.enable-integration=false
-quarkus.langchain4j.azure-openai.api-key=${AZURE_OPENAI_API_KEY:change-me}
-quarkus.langchain4j.azure-openai.embedding-model.enabled=false
-quarkus.langchain4j.azure-openai.moderation-model.enabled=false
-quarkus.langchain4j.azure-openai.image-model.model-name=gpt-image-2
-quarkus.langchain4j.azure-openai.image-model.quality=low
-quarkus.langchain4j.azure-openai.image-model.timeout=2m
-quarkus.langchain4j.azure-openai.resource-name=cs-super-heroes
-quarkus.langchain4j.azure-openai.deployment-name=csdeploy-super-heroes
-quarkus.langchain4j.azure-openai.timeout=15s
-quarkus.langchain4j.azure-openai.max-retries=3
-quarkus.langchain4j.azure-openai.chat-model.temperature=${quarkus.langchain4j.openai.chat-model.temperature}
-quarkus.langchain4j.azure-openai.chat-model.top-p=${quarkus.langchain4j.openai.chat-model.top-p}
-
 ## OpenAI profile configuration (in dev mode, make live calls to OpenAI by setting -Dquarkus.profile=openai)
 %dev,test,openai.quarkus.langchain4j.chat-model.provider=openai
 %dev,test,openai.quarkus.langchain4j.image-model.provider=openai
 %dev,test.quarkus.langchain4j.openai.base-url=http://localhost:${quarkus.wiremock.devservices.port}/v1/
 %dev,test.quarkus.langchain4j.openai.timeout=3s
-%dev,test,openai,azure-openai.quarkus.langchain4j.openai.enable-integration=true
-%dev,test,openai,azure-openai.quarkus.langchain4j.openai.log-requests=true
-%dev,test,openai,azure-openai.quarkus.langchain4j.openai.log-responses=true
+%dev,test,openai.quarkus.langchain4j.openai.enable-integration=true
+%dev,test,openai.quarkus.langchain4j.openai.log-requests=true
+%dev,test,openai.quarkus.langchain4j.openai.log-responses=true
 %dev,test.quarkus.langchain4j.openai.max-retries=2
 %openai.quarkus.langchain4j.openai.base-url=https://api.openai.com/v1/
 
-## Azure OpenAI Profile configuration (in dev mode, make live calls to Azure OpenAI by setting -Dquarkus.profile=azure-openai)
-%azure-openai.quarkus.langchain4j.chat-model.provider=azure-openai
-%azure-openai.quarkus.langchain4j.image-model.provider=azure-openai
-%dev,test.quarkus.langchain4j.azure-openai.endpoint=http://localhost:${quarkus.wiremock.devservices.port}/v1/
-%dev,test,quarkus.langchain4j.azure-openai.timeout=3s
-%dev,test,openai,azure-openai.quarkus.langchain4j.azure-openai.enable-integration=true
-%dev,test,openai,azure-openai.quarkus.langchain4j.azure-openai.log-requests=true
-%dev,test,openai,azure-openai.quarkus.langchain4j.azure-openai.log-responses=true
-%dev,test.quarkus.langchain4j.azure-openai.max-retries=2
-
 ## WireMock configuration
-# Disable it when making live calls to openai/azure openai
-%openai,azure-openai.quarkus.wiremock.devservices.enabled=false
+# Disable it when making live calls to openai
+%openai.quarkus.wiremock.devservices.enabled=false
 %dev,test.quarkus.wiremock.devservices.files-mapping=src/test/resources/wiremock
 
 ## HTTP configuration
@@ -69,7 +44,7 @@ quarkus.log.category."io.quarkus.sample.superheroes".level=DEBUG
 quarkus.log.level=INFO
 quarkus.log.console.level=INFO
 quarkus.log.console.darken=1
-%dev,test,openai,azure-openai.quarkus.log.console.level=DEBUG
+%dev,test,openai.quarkus.log.console.level=DEBUG
 
 # OpenTelemetry
 quarkus.otel.resource.attributes=app=${quarkus.application.name},application=narration-service,system=quarkus-super-heroes
@@ -83,7 +58,7 @@ quarkus.http.cors.origins=*
 
 # OpenAPI
 quarkus.smallrye-openapi.info-title=Narration API
-quarkus.smallrye-openapi.info-description=This API allows Narration using the Microsoft Semantic Kernel OpenAI
+quarkus.smallrye-openapi.info-description=This API allows Narration using OpenAI
 quarkus.smallrye-openapi.info-version=1.0
 quarkus.smallrye-openapi.info-contact-name=Quarkus
 quarkus.smallrye-openapi.info-contact-url=https://github.com/quarkusio
