--- rest-narration/src/test/java/io/quarkus/sample/superheroes/narration/rest/OpenAiNarrationResourceIT.java
@@ -1,139 +1,0 @@
-package io.quarkus.sample.superheroes.narration.rest;
-
-import static com.github.tomakehurst.wiremock.client.WireMock.*;
-
-import java.util.Map;
-
-import jakarta.ws.rs.core.HttpHeaders;
-import jakarta.ws.rs.core.MediaType;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.condition.DisabledIf;
-
-import io.quarkus.test.junit.QuarkusIntegrationTest;
-import io.quarkus.test.junit.QuarkusTestProfile;
-import io.quarkus.test.junit.TestProfile;
-
-import io.quarkus.sample.superheroes.narration.rest.OpenAiNarrationResourceIT.WiremockOpenAITestProfile;
-import io.quarkus.sample.superheroes.narration.service.ImageGenerationService;
-
-import io.quarkiverse.wiremock.devservice.WireMockConfigKey;
-
-@QuarkusIntegrationTest
-@TestProfile(WiremockOpenAITestProfile.class)
-@DisabledIf(value = "azureOpenAiEnabled", disabledReason = "Azure OpenAI profile is enabled")
-class OpenAiNarrationResourceIT extends NarrationResourceIT {
-	private static final String NARRATION_REQUEST_JSON = """
-    {
-      "model": "gpt-4o-mini",
-      "messages": [
-        {
-          "role": "system",
-          "content": "You are a marvel comics writer, expert in all sorts of super heroes and super villains."
-        },
-        {
-          "role": "user",
-          "content": "Narrate the matchup between a super hero and a super villain.\\n\\nDuring the narration, don't repeat \\"super hero\\" or \\"super villain\\".\\n\\nWrite 4 paragraphs maximum. Be creative and overly funny.\\n\\nThe narration must be:\\n- G rated\\n- Workplace/family safe\\n- No sexism, racism, or other bias/bigotry\\n- No overt violence. It needs to pass your own safety system and standards. Don't reference violence.\\n\\nHere is the data you will use for the winner:\\n\\n+++++\\nName: %s\\nPowers: %s\\nLevel: %d\\n+++++\\n\\nHere is the data you will use for the loser:\\n\\n+++++\\nName: %s\\nPowers: %s\\nLevel: %d\\n+++++\\n\\nHere is the data you will use for the fight:\\n\\n+++++\\n%s who is a %s has won the fight against %s who is a %s.\\n\\nThe fight took place in %s, which can be described as %s.\\n+++++\\n"
-        }
-      ],
-      "temperature": 1.0,
-      "top_p": 0.5,
-      "presence_penalty": 0.0,
-      "frequency_penalty": 0.0
-    }
-    """.formatted(
-    FIGHT.winnerName(),
-    FIGHT.winnerPowers(),
-    FIGHT.winnerLevel(),
-    FIGHT.loserName(),
-    FIGHT.loserPowers(),
-    FIGHT.loserLevel(),
-    FIGHT.winnerName(),
-    FIGHT.winnerTeam(),
-    FIGHT.loserName(),
-    FIGHT.loserTeam(),
-    FIGHT.location().name(),
-    FIGHT.location().description()
-  );
-
-  private static final String IMAGE_REQUEST_JSON = """
-    {
-      "model": "gpt-image-2",
-      "prompt": "%s\\nIn the gritty streets of Gotham City, a clash of epic proportions unfolded. Han Solo, a hero known for his sharpshooting skills and unwavering skepticism towards the force, faced off against Storm Trooper, a villain armed with nothing more than a small gun. The odds seemed stacked against the Storm Trooper, but he was determined to prove his worth.\\n\\nAs the battle commenced, Han Solo swiftly dodged the Storm Trooper's feeble shots, his agility and experience shining through. With a smirk on his face, Han Solo aimed his big gun with precision, firing shots that echoed through the city. The Storm Trooper, though outmatched, refused to back down, his determination fueling his every move.\\n\\nWith each passing moment, Han Solo's level of expertise became more apparent. His shots were calculated and deadly, while the Storm Trooper struggled to keep up. The hero's confidence grew, his movements becoming more fluid and effortless. It was clear that the Storm Trooper's small gun was no match for Han Solo's superior firepower.\\n\\nIn a final, decisive moment, Han Solo's shot found its mark, incapacitating the Storm Trooper. The hero emerged victorious, his unwavering resolve prevailing over the villain's futile attempts. As the city rejoiced in the triumph of justice, Han Solo stood tall, a symbol of hope and resilience in the face of adversity.\\nYou must answer strictly in the following JSON format: {\\n\\"url\\": (type: java.net.URI),\\n\\"base64Data\\": (type: string),\\n\\"mimeType\\": (type: string),\\n\\"revisedPrompt\\": (type: string)\\n}",
-      "n": 1,
-      "size": "1024x1024",
-      "quality": "low",
-      "moderation": "low"
-    }
-    """.formatted(ImageGenerationService.SYSTEM_MESSAGE.replaceAll("\n", "\\\\n"));
-
-  @Test
-	@Override
-	void shouldNarrateAFight() {
-		super.shouldNarrateAFight();
-
-		this.wireMock.verifyThat(
-      1,
-      postRequestedFor(urlPathEqualTo("/v1/chat/completions"))
-        .withHeader(HttpHeaders.ACCEPT, equalToIgnoreCase(MediaType.APPLICATION_JSON))
-        .withHeader(HttpHeaders.CONTENT_TYPE, equalToIgnoreCase(MediaType.APPLICATION_JSON))
-        .withHeader(HttpHeaders.AUTHORIZATION, equalTo("Bearer change-me"))
-        .withRequestBody(equalToJson(NARRATION_REQUEST_JSON, true, false))
-    );
-	}
-
-  @Test
-  @Override
-  void shouldGenerateAnImageFromNarration() {
-    super.shouldGenerateAnImageFromNarration();
-
-    this.wireMock.verifyThat(
-      1,
-      postRequestedFor(urlPathEqualTo("/v1/images/generations"))
-        .withHeader(HttpHeaders.ACCEPT, equalToIgnoreCase(MediaType.APPLICATION_JSON))
-        .withHeader(HttpHeaders.CONTENT_TYPE, equalToIgnoreCase(MediaType.APPLICATION_JSON))
-        .withHeader(HttpHeaders.AUTHORIZATION, equalTo("Bearer change-me"))
-        .withRequestBody(equalToJson(IMAGE_REQUEST_JSON, true, false))
-    );
-  }
-
-  @Test
-	@Override
-	void shouldGetAFallbackOnError() {
-		this.wireMock.register(
-      post(urlPathEqualTo("/v1/chat/completions"))
-        .withHeader(HttpHeaders.ACCEPT, equalToIgnoreCase(MediaType.APPLICATION_JSON))
-        .withHeader(HttpHeaders.CONTENT_TYPE, equalToIgnoreCase(MediaType.APPLICATION_JSON))
-        .withHeader(HttpHeaders.AUTHORIZATION, equalTo("Bearer change-me"))
-        .withRequestBody(equalToJson(NARRATION_REQUEST_JSON, true, false))
-        .willReturn(serverError())
-    );
-
-		super.shouldGetAFallbackOnError();
-
-		this.wireMock.verifyThat(
-      3,
-      postRequestedFor(urlPathEqualTo("/v1/chat/completions"))
-        .withHeader(HttpHeaders.ACCEPT, equalToIgnoreCase(MediaType.APPLICATION_JSON))
-        .withHeader(HttpHeaders.CONTENT_TYPE, equalToIgnoreCase(MediaType.APPLICATION_JSON))
-        .withHeader(HttpHeaders.AUTHORIZATION, equalTo("Bearer change-me"))
-        .withRequestBody(equalToJson(NARRATION_REQUEST_JSON, true, false))
-    );
-	}
-
-	public static class WiremockOpenAITestProfile implements QuarkusTestProfile {
-    @Override
-    public Map<String, String> getConfigOverrides() {
-	    var hostname = Boolean.getBoolean("quarkus.container-image.build") ? "host.docker.internal" : "localhost";
-
-      return Map.of(
-				"quarkus.langchain4j.openai.enable-integration", "true",
-        "quarkus.langchain4j.openai.log-requests", "true",
-        "quarkus.langchain4j.openai.log-responses", "true",
-        "quarkus.langchain4j.openai.base-url", "http://%s:${%s}/v1/".formatted(hostname, WireMockConfigKey.PORT),
-        "quarkus.langchain4j.openai.max-retries", "2",
-        "quarkus.langchain4j.openai.timeout", "3s"
-      );
-    }
-  }
-}
