+++ rest-narration/src/test/java/io/quarkus/sample/superheroes/narration/service/NarrationServiceDisabledTests.java
@@ -47,8 +47,7 @@ class NarrationServiceDisabledTests {
     @Override
     public Map<String, String> getConfigOverrides() {
       return Map.of(
-         "quarkus.langchain4j.openai.enable-integration", "false",
-         "quarkus.langchain4j.azure-openai.enable-integration", "false"
+         "quarkus.langchain4j.openai.enable-integration", "false"
       );
     }
   }
